pyproject.toml
[build-system]
requires = [
  "setuptools>=61.0",
  "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "mpesalens"
version = "0.1.0"
description = "MpesaLens - M-Pesa Statement Analyzer"
readme = "README.md"
requires-python = ">=3.8"
authors = [ { name = "Herman M. Omune", email = "hermanomune@gmail.com" } ]
license = { text = "MIT" }
keywords = ["mpesa", "streamlit", "finance", "analysis"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "streamlit>=1.0",
  "pandas>=1.5",
  "plotly>=5.0",
  "PyYAML>=6.0",
  "tabula-py>=2.0",
  "pikepdf>=7.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
